Add shortcode support and default data injection in Twig
Introduced the `wp_shortcode` function in `TwigWpExtension` to allow setting shortcodes when rendering with Twig. This enables the use of WordPress's built-in shortcodes for content generation. Also, enhanced the `TwigRenderer` by adding the `fill_default_data` function to auto-fill default WordPress data when rendering a template, improving code reusability and reducing redundancy.